Move InterfaceTarget to Target.ts
